Move record evaluation methods into data_evaluator module
